Refactor dir names
