+++ plugins/@grouparoo/vero/src/lib/export/exportArrayProperties.ts
